"2pm. Let me do the chores here.
2:10pm. I did not have to do much today. Okay, let me stop RI for the time being.

I've come to a decision, I am going to give up on input filtering. It seems like a nice optimization at first glance, but it has too many disadvantages when considered across various different contexts. For the GPU, rather than lowering its consumption when there is nothing to do, I really should be focusing on giving it more work instead.

The issues with contracting the input space as the processing goes along is not something I want to deal with.

2:30pm. Ok, enough. Stop taking breaks. Focus me. I can put a few hours into this.

2:50pm. My focus is high, but let me step away from the screen for a bit. I am converging to some conclussions, but need some time to elaborate them.

3:10pm. I took a short nap to think about it.

I've come to a decision - RNNs are out. There is no way to implement RNN agents efficiently on the GPU. It would be different if I had a specialized chip with its own dedicates threads for every outer dimension, but with the GPU, even for 2p games I'd have to waste 50% of the computation on every calculation.

This is because while p1 is acting, p2 will have to wait. With a feedforward net this would not be a problem since it has no state that needs to be carried over, but recurrent nets do. With 6p games I'd need to waste 5/6ths of the computational capacity!

The solution to make this efficient would be to use a data structure to store intermediate states for each player. But a regular matrix cannot be used for this, the data structure would have to be immutable. Or something like an array of arrays...

But either way, fetching these states from such a data structure in order to assemble them into a matrix would be expensive and complicated. And RNNs in particular are harder to train than feedforward nets.

Since transformers are feedforward nets, I should use them instead.

Also, when trained using backprop, RNNs are effectively feedforward anyway.

3:20pm. Forget it, I can safely leave RNNs out of consideration. Who knows if they will become dominant even on future hardware.

It is feedforward nets all the way for me from here. I should consider transformers for full NL holdem.

Hmmm, let me read that MLPmixer paper.

...Agh, let me reset the router.

https://arxiv.org/abs/2105.01601
MLP-Mixer - An all-MLP Architecture for Vision

I want to see what conclusions I can draw from this.

https://arxiv.org/abs/2007.07368
Explicit Regularisation in Gaussian Noise Injections

I might as well read this. I had the idea of injecting noise, and this should improve my understanding of what I was suggesting.

3:50pm. Ok, I think I get what MLP mixer is doing. It is interesting that it works so well.

Let me read the noise paper.

3:55pm. Ok, done. Some math stuff, not a big deal. Let me take a look at the MLP mixer again. What regularization is it using? It was not in the pseudocode.

> In particular, we use RandAugment [12], mixup [56], dropout [42], and stochastic depth [19].

Hadn't heard about the first two.

4:15pm. Done with lunch. It was a lot earlier today than I expected.

Mixup trains a net on convex combination of pairs of examples and their labels. Interesting.

4:20pm. Let me still my thoughts, they are too scattered around.

RNNs are out. Transformers, that is feedforward nets that can take in arbitrary length sequences will require some extra consideration for efficiency, but it is not a big deal.

4:30pm. The tabular methods, I'd also put into the feedforward category.

...Input filtering is out, just as I'd said earlier. Let me think through that again.

It is just so hard to start for me. Overcoming my internal inertia is a huge hurdle.

4:35pm. Let me step away from the screen.

6:50pm. Oh, it seems I stepped away for a while.

I was thinking about long term credit assignment and long term memory. Value functions and TD learning are a clue.

Now that I have my new training method, I can see gradient coming into the inputs as viable targets.

When composing networks which have different timescales, backprop would not work as by the time the slow net gets the targets, the fast net's trace would have already decayed.

The answer is to do what TD does - predict the targets. If it were before, that would have been absurd since the gradients could be badly unhinged, but once I rescale them, I'd get what I need.

Once the issue of passing rewards at different timescales becomes an issue, the only choice is to start predicting them. This is the answer, it cannot be anything else.

7pm. There was a period (probably in 2019) where I looked around for ways of making recurrence propagate info beyond the backprop horizon. There were some papers on that, but they were just hacks in the end. Fancy math, some benchmarks, but nothing that clicked.

I think this situation here is the same as with unsupervised learning. First there were autoencoders and they worked on just toys. Then there were GANs which worked on bigger toys, but couldn't possibly be the answer due to instability. Then there was the training scheme in the GMN paper that could make the stable.

Forget about unrolling RNNs for arbitrary number of steps. The only consideration RNNs should have is whether I want weight sharing across timesteps. I already know that LSTMs and regular RNNs cannot propagate info for very long distances regardless.

What can do it even with backprop are tree-like architectures in which the higher levels take info at increasingly longer timescales. Even if some trick exists that could make info propagate for a lot longer in RNNs are now, do I really think it will ever be the case that it will optimize for something like 2 ** 32 timesteps?

A tree NN could do that with 32 layers without trouble.

But even if the training could work in theory, the computer cannot possibly hold that many inputs in memory.

So the answer is the same as in TD learning - predict. When the memory is not big enough, use your own internal state as reference.

7:15pm. I do not know what the right scheme for doing that is, but it does not matter.

What matters is that I believe in it.

I've been agonizing about RNNs and architectures in the past, but that MLP Mixer paper proves that architectures are ultimately a technicality.

7:20pm. I should believe that everything I need is here. It will take a few more innovations to get where I want, but those will just be new spins on an old thing.

When truly long-term memory become necessary, the method to make synthetic gradients work will appear. For all I know maybe what exists today works well, though I doubt it.

7:30pm. Let me close here. My progress is snail-like, but tomorrow I will try to get the training procedure and the uniform agent to work as a sketch.

I am going to get through my inhibitions even if it kills me. Today I might not have done much programming, but I did not spent this time playing games or reading novels. I spent it trying to iron out my beliefs. Eventually the effort today will turn into motivation to do programming tomorrow. I will build the strength to travel the path.

I just had to think about the issue of the sequence length. It is inevitable that trying to decide between feedforward nets and recurrent ones would trigger this debate in me.

Not having a good reply for long term credit assignment is one of my weaknesses, but it is not that there aren't any hints in the present on what will the truth of tomorrow."