"10:10am. Let me chill a bit and then I will start. I want to study the two credit assignment papers by Dabney.
10:30am. Enough fun. Let me start.

https://arxiv.org/abs/2011.09464
Counterfactual Credit Assignment in Model-Free Reinforcement Learning

> In contrast, model-free reinforcement learning algorithms such as policy gradient methods (Williams, 1992; Sutton et al., 2000) perform simple time-based credit assignment, where events and rewards happening after an action are credited to that action, post hoc ergo propter hoc. While unbiased in expectation, this coarse-grained credit assignment typically has high variance, and the agent will require a large amount of experience to learn the correct relation between actions and rewards. Another issue of model-free methods is that counterfactual reasoning, i.e. reasoning about what would have happened had different actions been taken with everything else remaining the same, is not possible. Given a trajectory, model-free methods can in fact only learn about the actions that were actually taken to produce the data, and this limits the ability of the agent to learn quickly. As environments grow in complexity due to partial observability, scale, long time horizons, and large number of agents, actions taken by the agent will only affect a vanishing part of the outcome, making it increasingly difficult to learn from classical reinforcement learning algorithms. We need better credit assignment techniques

This seems pretty interesting. I wonder if it will turn out to be doing something better than what I had in mind. If that turns out to be the case, I should use it.

> CCA-PG also outperforms State- and ReturnConditional HCA, which do use hindsight information but in a more limited way than CCA-PG.

I guess I do not need that other paper.

10:55am. 4/32. I do not really understand how to calculate that hindsight quantity. Also how does it relate to the self probability?

I can see that it reweights the value estimate so that the current action probability in it is pushed towards something...maybe towards being uniform?

Actually, that does have implications on my own method. In my own method I do in fact reweight the Qs so the self reach probability is ommitted. But...that is a different thing from this.

11:35am. 25/32. I am mostly done with the paper, but this is difficult for me to understand. If I could work out how the update is derived from simpler models on my own, that would particularly help my understanding. It seems that this section is for that.

11:45am. > Interestingly, if we replace ∆i by the argmax of softmax of the ∆, we obtain a gradient estimate similar to that of the cross-entropy method, a classical and very effective planning algorithm (De Boer et al., 2005; Langlois et al., 2019).

I should look into the CEM at some point.

11:50am. I don't get a lot here, but I do get that they are making random numbers explicit in the Q estimate thereby making the entire model deterministic.

12pm. At any rate, now that I see this I can understand that...Ah, wait. There is variance from the the community card reveals. Knowing the opponent card does not give me minimized variance in poker.

True min variance is when everything is known by the Q function.

1:20pm. https://stats.stackexchange.com/questions/243015/why-do-variance-reduction-methods-work-well-and-are-there-any-backfire-experie

Done with breakfast. I want to study this more. I do not get the paper, but it is really making me think how future information can be used to do variance estimates. I am starting to get what the phi factor is, but in that case, what is the point of the reweighting? The paper is rather compilcated. I am wondering why not just shove all the future data into the Q function and call it a day?"