"1:45pm. Let me read Freiren.
The paper itself is really making me think about the importance of future information. If the paper is right and it is, then it is no wonder that OpenAI require batches of a million for their bots.

You absolutely need the hindsight information in real world tasks. The current techniques will not cut it. I do not actually need these techniques for poker, I can just hack the simulator to give the net all the info that it needs, but...I should understand this more in depth.

The subject of hindsight credit assignment is worth studying. So I should spend some time on this paper. Let's not hurry to start programming.

1:50pm. I really love this part of Freiren where she is autistically exploring every nook and cranny of a floor before moving on to the next one.

2:10pm. Yesterday it was interesting that both the transformer paper moved LN after the activation just before they got used as inputs. I can't remember what I did in the char RNN, but I think it was relu + LN rather than LN + relu. In the past I did not understand this, but now my knowledge should be enough to correctly interpret why that makes optimization easier.

My optimization scheme might make it unecessary, but it is always better safe than sorry. I'll try it out both ways with and without LN.

2:20pm. 28/32. Just what is this F here? I get the benefit of hindsight information, but how would you learn noise info? Like in that key door example, putting in future information would not have any effect if the number of apples is not given.

3:35pm. I think I get the paper in its entirety. That fi is just the hindsight state. It is nothing big at all.

The complications with the paper are maintaining the independence assumptions. They do a complicated thing which I barely get, but I can think of a better way.

What if instead of learning just `Q(H,A,F)`, the agent also learns `Q(H,A)`, that is without the hindsight information.

Then `W=Q(H,A,F) - Q(H,A)` could be considered the value of the hindsight info for the action taken.

Then the Qs for all the actions could be `Q'(H,A)=Q(H,A) + W`. This I could use for calculating the mean in tabular CFR methods. I am worried about the situations caused by taking actions that result in scenery switching which would cause the hindsight info to be completely irrelevant for those actions.

But the way the algorithm reweights that `P(H,A)` bothers me. They went all out to jack the standard PG training and made it quite complicated.

I wasn't going to, but let me look at the previous HCA paper.

4:15pm. Ok, I think I get a bunch of things now. I do not trust any of these algorithms. Their reweighting schemes are suspicious. As shady as regular PG is, the way it uses baselines is intuitive.

4:20pm. If I had to train an agent on a bigger game than poker, I'd use the variant I proposed above instead of any of the algos in these papers.

4:25pm. Forget hindsight methods. I won't have too much trouble dealing with them when the time comes.

No wait, that assumption I made that `Q'(H,A)=Q(H,A) + W` might not necessarily hold. With a different action, that W might flip signs or be worth something else...

How complicated. I'll throw this in the same pile as long term credit assignment. And speaking of that...

https://arxiv.org/abs/1810.06721
Optimizing Agent Behavior over Long Time Scales by Transporting Value

Let me take a break first. I've been focusing on these papers for a while.

4:35pm. I think that the standard way of doing thing would be the best. Forget W, just put the all the data right there even if there is a chance it might not have been trained on.

5:10pm. I am back. Let me read that paper I linked to. Forget about hindsight learning. It is not what I should be focusing on.

> In AI research, the problem of how to learn rational behavior that is temporally far-sighted is known as the credit assignment problem3–5. An AI agent must evaluate the utility of individual actions within a long sequence. To address the credit assignment problem, deep learning has been combined with reinforcement learning (RL) to provide a flexible class of architectures and algorithms that can be used practically to estimate the utility of courses of action for behaving agent models engaged in sensorimotor tasks in complex environments.

> These algorithms have almost exclusively borrowed the assumptions of discounted utility theory2, 6, 7 and achieve credit assignment using value function bootstrapping and backpropagation8 (deep RL). Practical and convergent deep RL algorithms discount the future, reducing their applicability for problems with long delays between decisions and consequences9, 10 .

> Conspicuously, humans and animals evidence behaviors that deep RL cannot yet simulate behaviorally. These come for example under the headings of latent learning11, 12, prospective memory13 , and inter-temporal choice14, and encompass learning and decision-making that occurs either without task reward, or when rewards are recouped at long delay from relevant choice points. It has been argued that hominid cognitive ability became truly modern when new strategies for long-term temporal credit assignment and planning emerged, leading to abrupt cultural shifts and immense changes in social complexity and human achievement15. Algorithmic progress on problems of long-term credit assignment may similarly lead to profound magnification of the range of decisionmaking problems that can be addressed computationally.

This paper was reference to in the HCA paper and it caught my eye. Long term credit assignment is one of my weak points. It is good that other people are working on it. I had no idea about those last few sentences. Without thinking about it, I just assumed that all animals had long term memory.

5:30pm. The paper is not much. It is just a hack test.

There isn't anything elegant on the market that combines hindsight, unsupervised learning and long term credit assignment to train RL agents. In fact, any of these individual 3 is hacky. There will be more on them.

I don't think that `Counterfactual Credit Assignment in Model-Free Reinforcement Learning` will be the last word on hindsight. Adding two extra nets that need optimizing is a huge burden and lacks elegance. Even of them even needs a strange constrained optimization algo.

5:40pm. What do I do here? I am satisfied with my efforts today. I do not feel like starting programming now.

5:45pm. It is difficult to come up with new ideas. Rather than using hindsight, for poker it is much smarter to just include the opponent's card. That is the way to go when doing variance reduction. For large games I'd just pass the inputs, including the future ones into the net as they are. No need for complicated stuff from these papers. Or my bad `W` idea. Whatever bias related degradation this would introduce will more than be made up by not needing to optimize two extra nets.

5:50pm. I'll admit, passing future inputs as they are is pretty nonsensical. What I'd really want is for them to be refined into something that would be useful regardless of the actions taken in the future trajectory. Eventually something more elegant will show up.

5:55pm. Now I am obsessing about that. Forget it. Tomorrow, I will do the uniform player and home in on doing the linear player. Future agents that do learning as well as humans can and have all the learning issues solved is something the future me can deal with.

Right now I a pauper so I should take the pennies tossed my way.

I absolutely need to focus. I need to do the uniform, linear, deep, and then the transformer agent. I need to get my things going. I had my fun with papers in the last two weeks, but now it is time to strike the iron and forge the blade. I'll do it a bit at a time.

6pm. Let me have some fun here. I feel like my mind has been in its normal state for the last few days. I should be able to program like this."