"10:10am. I've had difficulty falling asleep last night as my will is starting to concentrate. Stacking GANs independent as modules is my biggest hope for the future. I should be feeding them data and nurturing them.
The duality gap algo from the GMN paper for stabilizing their training is trustworthy. FID scores are sensitive to mode collapse so I should be able to trust them.

If it was before, stacking independent GAN modules would be suicide as their instability would compound, but if they are stable, they can form the foundation for an actual learning system.

If I passed them sequences, they would learn to predict the future.

The things I am missing for full AGI are some of the things the brain does. Maybe there will be different models better at pattern detection. Maybe those fast/slow weight ideas need revisiting.

But for the first time the foundations are now here. Long term memory formation can be induced with temporal bottlenecks and the saliency trick I've developed.

https://people.idsia.ch/~juergen/fast-weight-programmer-1991-transformer.html

Oh, he says that 2021 has new stuff.

https://arxiv.org/abs/2102.11174
Linear Transformers Are Secretly Fast Weight Memory Systems

Maybe attention is really all I need.

> Recent work proposed “linear Transformers” with constant size memory and time complexity linear in sequence length (Katharopoulos et al., 2020; Choromanski et al., 2021; Shen et al., 2018).

Huh they are not n^2 in sequence length? This is interesting!

> Therefore, we introduce an improved update rule inspired by recent work on fast weight memories (Schlag et al., 2021).

https://arxiv.org/abs/2011.07831
Learning associative inference using fast weight memory

Oh, Schlag is the first author of both papers.

10:50am. Had to take a break. For the morning my focus will be on these two papers. Without a doubt, the answer to the brain's short term memory questions is an architecture like this.

If I can provide an answer to the issues Simon Thorpe posed, I will have everything I need to scale my ideas to something much higher, even if the hardware is not there.

The computational insufficiency of the present does not matter. What matters is that my ideas will give me a clear path of advancement.

11:20am. The first paper was interesting, although propagating through the weights will blow up the memory requirements. Let me go for the second paper which uses LSTMs + fast weights. Something like that should be the answer.

> Lastly, we also evaluate Ba’s Fast Weights which attend to the recent past (JBFW; Ba et al. (2016a)) but were unable to find hyperparameters that converged.

Oh, lol.

> In our experiments, the LSTM-based agent requires more episodes, a bigger network, and eventually overfit to the training graphs. The FWM-based agent however trains faster and generalises to randomly sampled graphs.

> . The FWM agent (blue) has a slow LSTM with 32 hidden units and a fast weight memory of size 16 × 162 . We compare to LSTM agents with different sized hidden states. The largest LSTM has 4096 hidden units (red) which roughly matches the number of temporal variables of the FWM. The FWM has 14k trainable weights which is by far the lowest. The largest LSTM has 67.4M weights which is roughly 4814 times more than the FWM. The relative factor of each LSTM is added to the legend. All LSTMs take longer to train and eventually overfit on the training data. Due to the overfitting, the LSTM does not have to explore, which results in a higher total reward on training environments but a lower total reward on test environments.

If I am going to use an RNN, I should definitely use one that employs fast weights.

They use a fairly small weight matrix for the memories. That makes sense. They make a point that the human brain has a small working capacity as well.

11:45am. I don't know. It does not really matter at this stage.

Some kind of memory augmented RNN will come out in force in the future and become dominant, similar to how transformers are strong now. On neurochips, RNNs will be more efficient than feedforward nets whereas the story is the opposite on the GPUs.

11:50am. I am going to stop here. Enough of these papers. It is time for programming.

The way to advance in life is to practice agent cultivation. I might not have all the answers yet, but GAN training for prediction and my module ideas will stick around.

I am going to make a small RL agent that does even do unsupervised learning, but past that I am going to do my best to make agents that do. I will scale up. The modules I invented and the GAN duality gap method are what is going to make the whole thing scalable. I will believe in it.

The power of the agents will be my own power.

All I have to do to reach the answer at the end of the path is to build. It is time that I stop hesitating.

11:55am. I'll implement the uniform player after the breakfast."