"9:45am. I slept well tonight. I just need to go a bit further. I half about half of the true architecture in mind. Though I have the salience idea in mind, by itself it won't be enough to enforce temporal sparsity. But I do not want to just jam a bottleneck. I need one more step here, there should be some middle ground.
Also I need to bridge the gap in understanding on how to use unsupervised memory in NN layers. Transformers are supervised and they are great, but there should be more.

I won't be able to figure it out just like that, but I'll keep it in mind over the coming years. Right now, it does not even matter, since anything I could implement would just fry the GPU. I won't be able to move from simple games like poker for a while.

9:50am. Holy shit, Miura died. I thought it would be a joke when I saw it in the catalogue.

10:15am. Let me chill just a bit more.

10:30am. Let me start. I've slacked enough.

Time for the paper.

https://arxiv.org/abs/2008.02217
Hopfield Networks is All You Need

https://www.youtube.com/watch?v=IP3W7cI01VY
Lecture 11/16 : Hopfield nets and Boltzmann machines

Let me watch this for a bit. I want to remind myself of what Boltzmann machines are.

https://youtu.be/IP3W7cI01VY?t=2372

For the temporal bottleneck, it should be possible to smoothly interpolate between time horizons. The kind of bottleneck where each layer above is 2x slower might in theory be able to attent to exponentially long time horizons. But after a certain point, they'd never learn anything because the time horizons are simply too long. There has to be a better way of going about this. And just using saliency like I'd suggested would not be enough. It has to be a different kind of saliency.

10:50am. The video by Hinton is interesting, but back then they did not know what we know now.

https://youtu.be/IP3W7cI01VY?t=3002

Ah, I see. So Boltzman machines are quite similar to Hopfield nets. Visible are just units that have been clamped to some value.

https://youtu.be/IP3W7cI01VY?t=3078

Huh, this is the softmax. What are `u`, `g` all possible configurations? Ah, I see. In that case Boltzmann machines are not just an energy based model, but a probability distribution. Hopfield nets are not the later.

The first time I saw this video which was when it came out, I did not understand this at all, but now my understanding is much more vivid.

https://youtu.be/SY7ilsii2YM?t=16
Lecture 12/16 : Restricted Boltzmann machines (RBMs)

Let me watch this a little as well.

https://youtu.be/SY7ilsii2YM?t=390

Yeah, I understand now why I had so much difficulty understanding this back then. I did not understand the concept of sampling from a probability distribution. In fact, even now I am not sure how I'd sample from a Boltzmann machine. But it is not like these concepts are new to me. And now that I've read the EqProp thesis, I can relate that knowledge to this. The update rules for the weights of a Hopfield net and this are the same.

11:15am. How is sampling a RBM done? I forgot that.

11:30am. https://youtu.be/SY7ilsii2YM?t=1397

This is really vivid to me now. Back then I did not really understand how the weight updates for RBMs were derived. Probably because they were made up. But now there is a connection to EqProp.

https://youtu.be/SY7ilsii2YM?t=1787

I think I can anticipate how the sampling is done. I remember it iterating between forward and backward steps.

https://youtu.be/SY7ilsii2YM?t=2852

Yeah, I understand why you'd want to use a model that is an associative memory for this.

I am done with the video. I really get it completely now. Well, almost completely, I don't get how I'd implement MCMC for raw Boltzmann machines, but that does not matter.

11:45am. Let me go back to the paper.

3/95. So eq 3 here is for the reconstruction. Yeah, it is quite similar to the transformer rule. In the transformer X is calculated on the fly based on the keys. And big epsilon here is the query.

12pm. Ah, I thought that this might allow me to compress the data, but all the patterns need to be stored in a matrix as they are. Though, I guess that explains why they are exponential in capacity. It might be possible to reduce the store, by randomly sampling the matrix...

Ah, I see. If I have a replay buffer, I can easily turn that into a Boltzmann layer and sample from it.

12:15pm. Hmmm, is this really it. I would not have figured that the update rule given in the paper would work. Suppose you have a weight matrix of patterns like...

```
1,1,1
1,1,0
```

Then if you match a pattern like `1,1,0.1`, you'd expect it to be closer to the last one, but if a naive multiplication is done...

```
1*1+1*1+1*0.1=2.1
1*1+1*1+0*0.1=2
```

Then taking the softmax and summing the elements up, the reconstruction would be closer to the first element. I am not sure what to think about that. Simply because of that mental experiment, I'd assume that the algorithm is wrong.

Instead of multiplication it might be possible to fix it by taking the square distance. Or the absolute distance.

Hmmm, actually, suppose that 0.1 was negative. Then it would get further away from the first element.

So maybe it is fine. The noise would even out under this rule.

But the distance is the one which would get get glued to the first element even if it should not. The rule in the paper is the one which is in fact roboust.

12:30pm. Hmmm, one extra this paper gives me is a way of taking the weight matrix of a layer, and reconstructing its ideal input.

12:50pm. 11/95. Let me take a break here. I've read the body of the paper, the part starting here is the appendix.

So it turns out I did not need anything special to make the ideal associative single layer memories. The transformers were already it. When I saw the examples by Simon Thorpe, I was sure DL could not do it, but now I think it could given more capacity."