"1:25pm. Done with chores and breakfast, and the Mahoako chapter. I do not feel like programming right now as expected. It is really hard for me to get into the mindset.
Let me turn off the computer and step away from the screne for a while. I need to find the feeling again. Once I do, I will start this and continue going forward without stopping.

I am going to get the agent done, set it lose on online gambling dens and get my success and money. I will integrate new findings as they come along. I will follow the proper steps to get me to the main stage. But I need to get rid of the inertia and overcome the research mindset of the past month. It served me well for a while, but now it is time to go beyond.

Though it did not give me much in terms of algorithms, the Hopfield net paper and the research that I did improved my understanding of transformers significantly. For short term memory, attention is really all that I need.

There are some papers recently showcasing the benefits of feedforward nets, but transformers should have better generalization without the need for regularization. They will make better use of the GPU as well.

Eventually, the way to do Long Term Credit Assignment and proper modules will be discovered. My ideas are interesting pieces that I'll expect to play a role, but I can't see beyond them to the actual solution. I do not think the solution will be hard, I'd expect it to be simple, but regardless, I can't grasp it.

Once the hardware is there, and the understanding of the LTCA is complete, the Singularity itself will be close at hand.

The things I can't do right now do not matter. Ultimately, neuroscience will get its shit together. Or the researchers at large will stumble upon the solution.

I am envious and jealous, so much that I want to cry blood over not being able to go just a step further and deal with this myself. But in this competition, it is not like their position is that dominant. While they hide their thoughts, they are superior, but according to their incentives they have no choice but to release their insights, and once they are out, their work will be internalized by me.

They cannot hide their thoughts either because of the publish or perish culture. The only time hiding insight would make sense is if they stumble on something crucial while would enable them to dominate the whole world in a short amount of time. That kind of fantasy is not going to happen. The improvements are always incremental.

1:40pm. The last month was pretty fun. I learned quite a lot of ML that I did not know a few years ago. It has been a great refresher. But I need to put my own life in order.

I need to internalize the pursuit of the elite agent. The net architectures and the hardware will change, but the pipeline will stay the same. The refining the learning pipeline is something I can get better at and it will give me returns long into the future. The ML algorithms themselves are ultimately a small part of the overall scheme.

I will grow along with the technology itself.

Let me take a nap for a while here."