"7:40pm. Let me just add a note, if the credit assignment method that the brain uses is dependent on reconstruction as opposed to traces like backprop, then you'd expect it to want to make the inputs that have actually appeared more likely by lowering their energy. Otherwise reconstructions would be delusional.
How would lowering that energy manifest as?

```py
def E(s): -w * s[0] * s[1]
```

The derivative with respect to `w` is `-s[0] * s[1]`. `Input` times the `output` is the Hebbian learning rule.

The rule that is happening everywhere in the brain.

With backprop you don't need this as it has traces from the forward pass, but for long term credit assignment it would be essential."