"10:10am. I finished Doom Eternal again, but it was not satisfying even if it was fun. Compared to the way the DraQu and Zero Master play my own is just so clumsy. I can't immitate them. Their accuracy and awareness is on a higher level than mine, and their control is better. I think the later is because they are not using the two button mouse with a wheel like I am, but have something better. If I had something that allow me to use my right hand better, I could increase my efficiency. For example, being able to use the right thumb to change weapons would mean I do not have to take my left hand off the movement keys to do so. They way they swap between sniper and the balista indicates that they are doing something different from me.
10:15am. At any rate I am done with Doom again. Let me chill a bit and then I will start.

10:55am. Done with the break. Let me do some programming.

```
a = torch.rand(4,5)
ix = torch.tensor([0]).unsqueeze(-1)
print(a,torch.gather(a,1,ix))
print(ix.squeeze(-1).shape)
```

The way I was using squeeze without the dimension being given is wrong. 0 dimensional tensors are a thing and I do not want to get them by accident.

```
inl policy : a u64 f32 = $"torch.gather(!policy_probs_detached,1,!sample_indices.unsqueeze(-1)).squeeze(-1).numpy()"
```

This should do it.

```
    inl policy_raw_values' : obj = $"!policy.forward(!data_value[:,!policy_size])"
    inl policy_raw_values : obj = $"torch.full_like(!policy_raw_values',float('-inf'))"
    $"!policy_raw_values.flatten()[!action_indices] = !policy_raw_values'.flatten()[!action_indices]"
```

Oh, this is not good. Usually flatten will return a view, but since I am flattening a view, there is a high chance this will result in a copy.

Instead of setting it like this, I need to use a scatter.

https://pytorch.org/docs/stable/tensors.html#torch.Tensor.scatter_

```py
a = torch.zeros(4,5)
b = torch.ones(4,5)
ix = torch.tensor([[1,2],[2,3]])
a.scatter_(1,ix,b)
print(a)
```

This works great. This will make things a lot more efficient.

```
rm.add action_indices (am.map schema.actions.pickle actions)
```

Why is there a type error here?

```
    inl action_indices : ra u64 (a _ _) = am.empty
    l |> am.iteri fun b (player_state, (leduc_state (p1,p2,community_card)), pid, actions) =>
        inl seq = pl2_observations player_state pid |> obs_to_array
        inl pot = if pid = p1.id then p1.pot else p2.pot
        schema.value.pickle (leduc_policy_input {pid= $"!pid"; pot= $"!pot"; seq},if pid = p1.id then p2.card else p1.card) (0, $"!data_value[!b,:]")
        rm.add action_indices (am.map schema.actions.pickle actions)
```

Because of that type annotation. This passes, but it is not right. In order to convert these lists into a tensor the dimensions have to be matching.

What a bother. If I wanted to fix this, I would not even need the scatter I could just do a mask.

11:20am. I am going to need to do the same thing for the reward on the backward pass. Then it is settled, let me do a mask.

```
    inl data_value : obj = $"torch.zeros(!len,!value_size)"
    l |> am.iteri fun b (player_state, (leduc_state (p1,p2,community_card)), pid, actions) =>
        inl seq = pl2_observations player_state pid |> obs_to_array
        inl pot = if pid = p1.id then p1.pot else p2.pot
        schema.value.pickle (leduc_policy_input {pid= $"!pid"; pot= $"!pot"; seq},if pid = p1.id then p2.card else p1.card) (0, $"!data_value[!b,:]")
    inl action_indices : a u64 (a u64 _) = l |> am.generic.map fun (_,_,_,actions) => am.map schema.actions.pickle actions
    inl policy_mask : obj = $"torch.full((!len,!actions_size),float('-inf'))"
    am.iteri (fun b => am.iter (fun a => $"!policy_mask[!b,!a] = 0")) action_indices
    inl policy_raw_values : obj = $"!policy_mask + !policy.forward(!data_value[:,!policy_size])"
    inl policy_probs : obj = $"torch.softmax(!policy_raw_values,dim=-1)"
```

Yeah, this is good.

https://pytorch.org/docs/stable/tensors.html#torch.Tensor.cpu

It does not do a copy if the tensor is already on the cpu. Ok.

12:30pm. Let me stop here since I've already started lurking the isekai thread on the side.

```
    inl sample_indices : a u64 _ = $"torch.distributions.Categorical(!e + (1.0 - !epsilon) * !policy_probs_detached).sample().numpy()"
    inl policy : a u64 f32 = $"torch.gather(!policy_probs_detached,1,!sample_indices.unsqueeze(-1)).squeeze(-1).numpy()"
    inl l : a _ _ =
        am.map2 (fun policy a =>
            inl policy = f64 policy
            inl sample : f64 = e + (1.0 - epsilon) * policy
            log_prob_from {policy sample}, schema.actions.unpickle a
            ) policy sample_indices
    l, fun (r : a u64 r2) =>
        inl value_raw : obj = $"!value.forward(!data_value)"
        // inl value : a u64 f32 =
        r
```

It is a huge pain in the ass for me to decide how I want to do the CFR update here.

Gather and scatter have reduce operations. But to make use of that, can I do something like zipping a tensor to be operated upon first?

I need to combine those `r`s with the policy probability.

How annoying. Let me stop here. I'll figure this out eventually."